Implement tests for LUT decoder.
