+++ test/main.cpp
@@ -2,16 +2,15 @@
 #include <vector>
 
 #include "mortonND_test.h"
-#include "mortonND_LUT_encoder_test.h"
-#include "mortonND_LUT_decoder_test.h"
+#include "mortonND_LUT_test.h"
 #include "mortonND_BMI2_test.h"
 
 #include <iostream>
 
 auto test_methods = std::vector<test_method>{
-    //test_method(&mortonnd_lut::TestEncode, "Test LUT encoder configurations (dimension, field size, LUT entry size)."),
-    //test_method(&mortonnd_bmi2::TestEncode, "Test BMI2 encoder configurations (dimension, field size)."),
-    //test_method(&mortonnd_bmi2::TestDecode, "Test BMI2 decoder configurations (dimension, field size).")
+    test_method(&mortonnd_lut::TestEncode, "Test LUT encoder configurations (dimension, field size, LUT entry size)."),
+    test_method(&mortonnd_bmi2::TestEncode, "Test BMI2 encoder configurations (dimension, field size)."),
+    test_method(&mortonnd_bmi2::TestDecode, "Test BMI2 decoder configurations (dimension, field size)."),
     test_method(&mortonnd_lut::TestDecode, "Test LUT decoder configurations (dimension, field size, LUT entry size).")
 };
