--- test/mortonND_LUT_decoder_test.h
@@ -1,5 +1,0 @@
-#pragma once
-
-namespace mortonnd_lut {
-    bool TestDecode();
-}
