+++ test/mortonND_LUT_test.h
@@ -2,4 +2,5 @@
 
 namespace mortonnd_lut {
 bool TestEncode();
+bool TestDecode();
 }
